+++ Database/DTOs/PricesDTO/BaseIdentityPricesDTO.cs
@@ -3,7 +3,7 @@ namespace Database.DTOs.PricesDTO
 {
     public class BaseIdentityPricesDTO : BasePricesDTO
     {
-        public BaseIdentityPricesDTO(int prcid, int plid, string dps, string dpe, float price, bool ispt, string days) 
+        public BaseIdentityPricesDTO(int prcid, int plid, string dps, string dpe, decimal price, bool ispt, string days) 
             : base(plid, dps, dpe, price, ispt, days)
         {
             this.prcid = prcid;
