+++ Database/DTOs/PricesDTO/UpdatePricesDTO.cs
@@ -3,7 +3,7 @@ namespace Database.DTOs.PricesDTO
 {
     public class UpdatePricesDTO : BaseIdentityPricesDTO
     {
-        public UpdatePricesDTO(int prcid, int plid, string dps, string dpe, float price, bool ispt, string days) 
+        public UpdatePricesDTO(int prcid, int plid, string dps, string dpe, decimal price, bool ispt, string days) 
             : base(prcid, plid, dps, dpe, price, ispt, days)
         {
         }
