+++ Database/DTOs/SeasonalitiesDTO/BaseIdentitySeasonalitiesDTO.cs
@@ -4,7 +4,7 @@ namespace Database.DTOs.SeasonalitiesDTO
 {
     public class BaseIdentitySeasonalitiesDTO
     {
-        public BaseIdentitySeasonalitiesDTO(int seasid, string stdt, string endt, double coef)
+        public BaseIdentitySeasonalitiesDTO(int seasid, string stdt, string endt, decimal coef)
         {
             this.seasid = seasid;
             this.stdt = stdt ?? throw new ArgumentNullException(nameof(stdt));
@@ -15,6 +15,6 @@ namespace Database.DTOs.SeasonalitiesDTO
         public int seasid { get; set; }
         public string stdt { get; set; }
         public string endt { get; set; }
-        public double coef { get; set; }
+        public decimal coef { get; set; }
     }
 }
