+++ Database/DTOs/SeasonalitiesDTO/CreateSeasonalitiesDTO.cs
@@ -3,7 +3,7 @@ namespace Database.DTOs.SeasonalitiesDTO
 {
     public class CreateSeasonalitiesDTO : BaseIdentitySeasonalitiesDTO
     {
-        public CreateSeasonalitiesDTO(int seasid, string stdt, string endt, double coef) 
+        public CreateSeasonalitiesDTO(int seasid, string stdt, string endt, decimal coef) 
             : base(seasid, stdt, endt, coef)
         {
         }
