+++ Database/DTOs/SeasonalitiesDTO/SeasonalitiesDTO.cs
@@ -4,7 +4,7 @@ namespace Database.DTOs.SeasonalitiesDTO
     public class SeasonalitiesDTO : BaseIdentitySeasonalitiesDTO
     {
 
-        public SeasonalitiesDTO(int seasid, string stdt, string endt, double coef) 
+        public SeasonalitiesDTO(int seasid, string stdt, string endt, decimal coef) 
             : base(seasid, stdt, endt, coef)
         {
         }
