+++ Database/DTOs/SeasonalitiesDTO/UpdateSeasonalitiesDTO.cs
@@ -3,7 +3,7 @@ namespace Database.DTOs.SeasonalitiesDTO
 {
     public class UpdateSeasonalitiesDTO : BaseIdentitySeasonalitiesDTO
     {
-        public UpdateSeasonalitiesDTO(int seasid, string stdt, string endt, double coef) 
+        public UpdateSeasonalitiesDTO(int seasid, string stdt, string endt, decimal coef) 
             : base(seasid, stdt, endt, coef)
         {
         }
