+++ Database/DTOs/SectablesDTO/UpdateSectablesDTO.cs
@@ -3,7 +3,7 @@ namespace Database.DTOs.SectablesDTO
 {
     public class UpdateSectablesDTO : BaseIdentitySectablesDTO
     {
-        public UpdateSectablesDTO(int sctid, int sec, double coef) 
+        public UpdateSectablesDTO(int sctid, int sec, decimal coef) 
             : base(sctid, sec, coef)
         {
         }
